Packed QuarterlyDistribution struct

+++ src/grants/base/StandardFunding.sol
@@ -202,11 +202,11 @@ abstract contract StandardFunding is Funding, IStandardFunding {
 
         // create QuarterlyDistribution struct
         QuarterlyDistribution storage newDistributionPeriod = distributions[newDistributionId_];
-        newDistributionPeriod.id              = newDistributionId_;
-        newDistributionPeriod.startBlock      = startBlock;
-        newDistributionPeriod.endBlock        = endBlock;
+        newDistributionPeriod.id              = uint128(newDistributionId_);
+        newDistributionPeriod.startBlock      = uint128(startBlock);
+        newDistributionPeriod.endBlock        = uint128(endBlock);
         uint256 gbc                           = Maths.wmul(treasury, GLOBAL_BUDGET_CONSTRAINT);  
-        newDistributionPeriod.fundsAvailable  = gbc;
+        newDistributionPeriod.fundsAvailable  = uint128(gbc);
 
         // decrease the treasury by the amount that is held for allocation in the new distribution period
         treasury -= gbc;
